Added support for embedded SVG
Pandoc's previous behavior was to look for a "blip" element.

Microsoft has introduced an extension offering the "svgBlip" element to
support SVG graphics. Whenever a document contains SVG it also contains
a PNG for backwards compatibility.

These changes are modest:

(1) look for an svgBlip *first*. If found, include that.
(2) if not found, fallback to blip
(3) I put in some placeholder code for "link" pictures. This code
    seems to not do any harm because the rid won't resolve without
    some further work (next commit)